#55 Refactor Form API
