+++ extension/ts/extensions/pdf.ts
@@ -1,7 +1,7 @@
 import {showToast, ToastType} from "../ui/toast";
 import {createLoader, removeLoader} from "../ui/loadingIndicator";
-import {FORM_RENDER_EVENT} from "../services/renderFormObserver";
 import {find} from "../services/finder/finder";
+import {onFormRender} from "../objects/bookForm";
 
 const findTextContent = (id: string) => (): string =>
 	(document.getElementById(id) as HTMLTextAreaElement | HTMLInputElement)?.value ?? "";
@@ -36,7 +36,7 @@ const createButton = (onClick: (event: MouseEvent) => void): HTMLButtonElement =
 	return button;
 };
 
-window.addEventListener(FORM_RENDER_EVENT, async () => {
+onFormRender(() => {
 	const commentsCell = document.getElementById("bookedit_comments");
 	const comments = document.getElementById("form_comments") as HTMLTextAreaElement; // not type safe -- lazy
